Move promotion to reusable component
